requirements.txt
cycler==0.10.0
dataclasses==0.6
future==0.18.2
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
threadpoolctl==2.1.0
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.0
tqdm==4.61.0
typing-extensions==3.10.0.0
